requirements.txt
openai-whisper
openai
sounddevice
pyaudio
pyautogui
pygetwindow
opencv-python
numpy